+++ LAGGED/src/Utility/Noise.cpp
@@ -11,7 +11,7 @@ namespace LAG::Noise
 		glm::vec2 scale = size + 1.f;
 
 		noiseData.m_Size = scale;
-		noiseData.m_Data.resize((scale.x) * (scale.y));
+		noiseData.m_Data.resize(static_cast<size_t>((scale.x) * (scale.y)));
 
 		if (size.x > 0 && size.y > 0)
 		{
@@ -19,7 +19,7 @@ namespace LAG::Noise
 
 			//TODO: generator should be stored
 			auto generator = FastNoise::NewFromEncodedNodeTree("GQANAAMAAACamRk/GwANAAIAAAC4HoU+BwAAuB4lQACF61E/AAAAgEAAAAAAQQAAAAAAASQAAgAAABwAARkADQACAAAAuB4FQP//AAAAj8I1QQB7FK4+AQ0AAgAAAMP1qED//wAAAJqZmT4AcT0KPwApXE9A");
-			generator->GenUniformGrid2D(noiseData.m_Data.data(), static_cast<int>(position.x), static_cast<int>(position.y), scale.x, scale.y, noiseProps.m_Frequency * frequencyAdjustment, noiseProps.m_Seed);
+			generator->GenUniformGrid2D(noiseData.m_Data.data(), static_cast<int>(position.x), static_cast<int>(position.y), static_cast<int>(scale.x), static_cast<int>(scale.y), noiseProps.m_Frequency * frequencyAdjustment, noiseProps.m_Seed);
 			if (noiseProps.m_Amplitude != 1.f)
 			{
 				for (int i = 0; i < noiseData.m_Data.size(); i++)
@@ -37,7 +37,7 @@ namespace LAG::Noise
 		GeneratedData noiseData = GenerateNoiseData(noiseProperties, position, size, surfaceSubdivision);
 		Texture texture(HashedString(""));
 
-		texture.SetBuffer(noiseData.m_Data.data(), noiseData.m_Data.size(), noiseData.m_Size.x, noiseData.m_Size.y, LAG::TextureFormat::FORMAT_R);
+		texture.SetBuffer(noiseData.m_Data.data(), noiseData.m_Data.size(), static_cast<int>(noiseData.m_Size.x), static_cast<int>(noiseData.m_Size.y), LAG::TextureFormat::FORMAT_R);
 		texture.Reload();
 
 		return texture;
